Fix return value of FileMover.call
Update FileMover.call, which was not completely integrated with previous change that switched things from being purely boolean, to having a range of possible statuses.  The method was setting a local variable to false, and always returning that.  It turns out that the return value does not matter for much.  But have it return the right one, anyway.

Also make sure the status is failure upon an exception, and update the Javadoc.