Back out explicit MoveStatus
I just recently added a MoveStatus top-level enum.  But now I'm backing it out, in favor of a series of boolean variables kept in the FileEpisode.  Rather than setting the status to a different enum, the FileMover will now call a similarly named, void method to "update" the status.  Move the formerly private method FileMover.isSuccess, into a public method in FileEpisode.

Change finishProgress and finishMove to take the FileEpisode rather than just a boolean that indicated success.