Use finally clause to always call finishProgress
It used to be that the only reason to use "finishProgress" was to finish the updates that were being made with setProgressValue, which meant it only needed to be called if we were doing a copy-and-delete.  But now we use "finishProgress" to check the success or failure of the overall operation, so we can inform the user.  So we should call it every single time.  Move it into the call() method, in a finally clause, to try to make sure it gets called once per FileMover, no matter what.