Change documentation around MoveObserver
Previously, the MoveObserver (formerly ProgressObserver) was used only in the case of a "copy and delete", in order to update the graphical progress indicator.  But now we use it to know when ANY move has completed, whether a simple rename or a copy, and whatever the level of success.

It used to be true that if you weren't going to send progress updates, you didn't need to touch the observer at all.  You didn't initialize it, nor "finish" it.

It is still true that you don't need to call the explicit initialization method unless you intend to do progress updates.  Have the doc reflect that.

But it's also now true that if you have an observer, when you are done you should always call finishProgress.  Our initial change to MoveObserver did not do this, but the previous commit did.  Now this commit just updates comments, and also changes the name (no longer File "Copy" Monitor, but just FileMonitor) to reflect the new usage.