Delete some nonfunctional MoveRunner stuff
This has to do with conflicts.

Let's say we have two files: /Users/me/Incoming/Lost.6x05.avi, and /Users/me/Incoming/Lost.S06E05.avi

After looking up the information, and filling in the replacement template, it turns out we want to move them both to the same location (let's say /Users/me/TV/Lost/Season06/Lost.S06E05.Lighthouse.avi).

We want to detect the conflict, and not lose either file.  There really are numerous options of what to do, and we probably should expose more to the user, to let them decide.

For now, what we do is we try to move one to the desired name, and the other to a different name.

For information, what we try to move the other one to is /Users/me/TV/Lost/Season06/versions/Lost.S06E05.Lighthouse (2).avi

(That is, we will add a "(2)" to the filename, before the extension, and we will ALSO move it into a subdirectory called "versions".)

But those specifics are not actually important right now.

Now let's say we're in a slightly different situation.  The only file we have to move is /Users/me/Incoming/Lost.6x05.avi.  There are no conflicts among the moves we want to execute.

However, what if the destination file (/Users/me/TV/Lost/Season06/Lost.S06E05.Lighthouse.avi) already exists?

That's a conflict just like before.  We do the same thing: add the (2) and put it in "versions".

Now, let's say we have BOTH situations.  We have two moves going to the identical place, and the destination file is already there.  Well, we handle that, too.  You guessed it, the incoming files get moved to (2) and (3).

Still, we haven't covered everything.  What if the alternate file, the one with "(2)", already exists, too?

The answer is that, as of the previous version, and everything following it, we fail to move the file.  Fortunately, we don't overwrite anything, so we don't lose any data, but we do not detect the conflict, and therefore we fail.

I originally wrote this in a way that would not fail, because it would detect all kinds of potential conflicts, by looking only at the "basename" of the desired destination.  That is, I'd look at "Lost.S06E05.Lighthouse".  Beneath the destination directory, I'd look for ANY file that began with that string.

I did that somewhat purposefully.  I didn't particularly care if it was "x (2).avi" or "x-2.avi" or "x backup.avi"; I wanted to just have the number reflect how many files appeared to refer to the same episode.

And, I very much intended, at the time, to ignore the filename extension.  That is, "Lost.S06E05.Lighthouse.avi" and "Lost.S06E05.Lighthouse.mkv" are very much a conflict.  If they refer to the same episode, it doesn't matter which video container format they're using.  Even though they wouldn't overwrite each other, I wanted it called out as a duplicate.

The problem, that I didn't think of, was that this also considered it a conflict to have "Lost.S06E05.Lighthouse.avi" and "Lost.S06E05.Lighthouse.srt".  One is the primary file, the movie file, and the other is an auxiliary file, the subtitles.  They are not conflicts; they are complementary and should be kept together.

Since I released a version with this bug, I wanted to fix it as quickly as I could; so, instead of looking at just the basename ("Lost.S06E05.Lighthouse"), I looked at the entire desired filename ("Lost.S06E05.Lighthouse.avi").

That solved the problem of the auxiliary files being considered a conflict, but it led to the problem where we don't detect alternate files being in the way.

(It also causes us to not see ".avi" vs. ".mkv" as a conflict, but that's not relevant to this commit.)

This commit does not fix things.  All it does do is, clean up some of the cruft that was left behind from the first, abortive attempt, and make it a little more like what the code should be like if I had written it this way from the beginning.

We stop doing filename globbing, or searching the directory for a conflicting file.  If there's a conflict, we know exactly what it will be called, so simply look for that file.  Update the doc and the variable names to know we're not actually dealing with a "basename" (which implies no file extension), but a filename.

I intend to get back to this, and fix the existing bug.  But clean things up first.