Modify FileMover method into versionString
We used to have private method addVersionString, which concatenated the basename, the index string (which it created), and the suffix.  Change it to just create the index string, and make it package-private.  Its one caller in FileMover.java will now be responsible for doing the concatenation.

The main reason for this change at this time is to allow test classes to access the functionality.  I also do want to make the naming configurable, in the future (#267).